requirements.txt
aiofiles==24.1.0
aiogram==3.20.0.post0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
attrs==25.3.0
certifi==2025.4.26
dnspython==2.7.0
frozenlist==1.6.0
idna==3.10
magic-filter==1.0.12
multidict==6.4.3
propcache==0.3.1
psutil==7.0.0
pydantic==2.11.4
pydantic-settings==2.9.1
pydantic_core==2.33.2
python-dotenv==1.1.0
setuptools==80.4.0
typing-inspection==0.4.0
typing_extensions==4.13.2
yarl==1.20.0


mistralai~=1.7.0
chromadb~=1.0.9
Pyrogram~=2.0.106
Deprecated~=1.2.18
sentence-transformers~=4.1.0
openai~=1.78.1
nltk
emoji
TgCrypto

sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.13.1
psycopg2-binary